Upgrade SlidePane and Dialog to use classes
