+++ src/dialog/styles/dialog.css.d.ts
@@ -1,5 +1,6 @@
 export const main: string;
 export const underlay: string;
+export const underlayVisible: string;
 export const title: string;
 export const content: string;
 export const close: string;
