+++ src/slidepane/example/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-	<title>SlidePanel Example</title>
+	<title>SlidePane Example</title>
 	<meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1">
 </head>
 <body>
@@ -26,7 +26,7 @@
 		});
 
 		/* Requiring in the main module */
-		require([ 'src/slidePanel/example/index' ], function () {});
+		require([ 'src/slidepane/example/index' ], function () {});
 	</script>
 </body>
 </html>
