+++ src/slidepane/example/index.ts
@@ -1,78 +1,69 @@
-import { DNode, Widget, WidgetProperties } from '@dojo/widget-core/interfaces';
-import { w, v } from '@dojo/widget-core/d';
-import createWidgetBase from '@dojo/widget-core/createWidgetBase';
-import createProjectorMixin from '@dojo/widget-core/mixins/createProjectorMixin';
-import createSlidePanel, { Align } from '../../slidePanel/createSlidePanel';
-import internalState, { InternalState } from '@dojo/widget-core/mixins/internalState';
+import { WidgetBase, WidgetProperties } from '@dojo/widget-core/WidgetBase';
+import { StatefulMixin } from '@dojo/widget-core/mixins/Stateful';
+import { ProjectorMixin } from '@dojo/widget-core/mixins/Projector';
+import { v, w } from '@dojo/widget-core/d';
+import SlidePane, { Align } from '../../slidepane/SlidePane';
 
-type Root = Widget<WidgetProperties> & InternalState;
-
-function openSlidePanel(this: Root) {
-	this.setState({ open: true });
-}
-
-function toggleUnderlay(this: Root, event: Event) {
-	this.setState({ underlay: (<HTMLInputElement> event.target).checked });
-}
-
-function toggleAlign(this: Root, event: Event) {
-	this.setState({ align: (<HTMLInputElement> event.target).checked ? Align.right : Align.left });
-}
-
-const createApp = createWidgetBase.mixin(internalState).mixin({
-	mixin: {
-		cssTransitions: true,
-		getChildrenNodes: function(this: Root): DNode[] {
-			return [
-				w(createSlidePanel, {
-					id: 'panel',
-					open: <boolean> this.state['open'],
-					underlay: <boolean> this.state['underlay'],
-					align: <Align> this.state['align'],
-					onRequestClose: () => {
-						this.setState({ open: false });
-					}
-				}, [
-					`Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-					Quisque id purus ipsum. Aenean ac purus purus.
-					Nam sollicitudin varius augue, sed lacinia felis tempor in.`
-				]),
-				v('button', {
-					id: 'button',
-					innerHTML: 'open slidepanel',
-					onclick: openSlidePanel
-				}),
-				v('div', { classes: { option: true }}, [
-					v('input', {
-						type: 'checkbox',
-						id: 'underlay',
-						onchange: toggleUnderlay
-					}),
-					v('label', {
-						for: 'underlay',
-						innerHTML: 'underlay'
-					})
-				]),
-				v('div', { classes: { option: true }}, [
-					v('input', {
-						type: 'checkbox',
-						id: 'alignRight',
-						onchange: toggleAlign
-					}),
-					v('label', {
-						for: 'alignRight',
-						innerHTML: 'align right'
-					})
-				])
-			];
-		},
-		classes: [ 'main-app' ],
-		tagName: 'main'
+export class App extends StatefulMixin(WidgetBase)<WidgetProperties> {
+	openSlidePane() {
+		this.setState({ open: true });
 	}
-});
 
-createApp.mixin(createProjectorMixin)({
-	cssTransitions: true
-}).append().then(() => {
-	console.log('projector is attached');
-});
+	toggleUnderlay(event: Event) {
+		this.setState({ underlay: (<HTMLInputElement> event.target).checked });
+	}
+
+	toggleAlign(event: Event) {
+		this.setState({ align: (<HTMLInputElement> event.target).checked ? Align.right : Align.left });
+	}
+
+	render() {
+		return v('div', [
+			w(SlidePane, {
+				id: 'pane',
+				open: <boolean> this.state['open'],
+				underlay: <boolean> this.state['underlay'],
+				align: <Align> this.state['align'],
+				onRequestClose: () => {
+					this.setState({ open: false });
+				}
+			}, [
+				`Lorem ipsum dolor sit amet, consectetur adipiscing elit.
+				Quisque id purus ipsum. Aenean ac purus purus.
+				Nam sollicitudin varius augue, sed lacinia felis tempor in.`
+			]),
+			v('button', {
+				id: 'button',
+				innerHTML: 'open slidepane',
+				onclick: this.openSlidePane
+			}),
+			v('div', { classes: { option: true }}, [
+				v('input', {
+					type: 'checkbox',
+					id: 'underlay',
+					onchange: this.toggleUnderlay
+				}),
+				v('label', {
+					for: 'underlay',
+					innerHTML: 'underlay'
+				})
+			]),
+			v('div', { classes: { option: true }}, [
+				v('input', {
+					type: 'checkbox',
+					id: 'alignRight',
+					onchange: this.toggleAlign
+				}),
+				v('label', {
+					for: 'alignRight',
+					innerHTML: 'align right'
+				})
+			])
+		]);
+	}
+}
+
+const Projector = ProjectorMixin(App);
+const projector = new Projector({});
+
+projector.append();
