+++ src/styles/widgets.css
@@ -1,3 +1,3 @@
 @import 'animations.css';
 @import '../dialog/styles/dialog.css';
-@import '../slidePanel/styles/slidePanel.css';
+@import '../slidepane/styles/slidePane.css';
